requirements-dev.txt
# requirements-dev.txt
-e .
black==24.2.0
isort==5.13.2
mypy==1.9.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dotenv
